HISTORIAL MEDICO CON FILTROS FUNCIONAL
YA FUNCIONA LA VISTA DE HISTORIAL MEDICO CON TODOS LOS DETALLES Y SUS FILTROS

+++ MediSys/Models/User.cs
@@ -8,7 +8,7 @@ namespace MediSys.Models
 		public int IdUsuario { get; set; }
 
 		[JsonPropertyName("cedula")]
-		public long Cedula { get; set; } // 🔥 CAMBIO: long en lugar de string
+		public int Cedula { get; set; } // 🔥 CAMBIO: long en lugar de string
 
 		[JsonPropertyName("username")]
 		public string Username { get; set; } = string.Empty;
